release.sh script to build and push riffmm and edu-web; & web/server ver map
release.sh was copied from riff-docker and I added new support for riffmm and
edu-web but did not remove any targets (even those not really appropriate for
edu) because a unified script should be easier to maintain, and perhaps there
will be a way to run it from a single location eventually.

Also riffmm-versions is where the map of the riffmm version (specified as the
-r argument to release.sh) to the source version of mm-web and mm-server that
are used by that riffmm version is kept (and updated)
I've initialized riffmm-versions with the versions we've deployed up to now.

Currently edu-web doesn't really change so it's just being tagged as latest
ie
bin/release.sh -r latest build edu-web
bin/release.sh -r latest push edu-web