change 'make up' to use released versions of riffdata and signalmaster
docker-compose.yml defaults to "latest" for riffdata which is a tag we have not yet started applying so the dev config sets a specific release version.

I made this change because it turns out that we pretty much never need a development version of riffdata and signalmaster AND because I moved the riffedu specific repos to a different directory from the riffplatform repos so ../riff-server and ../signalmaster is not the same location from edu-docker as from riff-docker.