publish the riffdata container port 3000
This is so that it may be accessed by another AWS instance in the same VPC over the internal network.

In particular a jitsi meet instance will share the riffdata of this riffedu instance so that it can send utterance data and get turn and participant data during a meeting, and the riffedu instance can get the meeting data to display metrics.