build: move docker-compose depends_on config to a separate config file
This was done to enable the Makefile's new target `dev-mm-min` which starts ONLY the edu-mm service AND NOT the edu-mm-db service. This will allow interactively building the riff mm server, webapp and plugins, but not running any tests that require the database.

The reason this became an issue is a developer using a Mac w/ an ARM64 cpu, and there is no supporting image of mysql:5.7 for that CPU architecture.